Phase Two (2) Revisions (#75)
* Fixed the issues about the logo upload and print.

* Exployee module place of birth label note format.

* Changes in logging exception errors in controller.
Added a general function for reusability.

* Removedclient fields in view and dispatch table.

* Inventory and PRF revisions:
- Delears Price to Item Cost
- Added date format in PRF # - YYYYMMDD-ID
- Process Date to Date Needed
- Removed LESS THAN STOCK validation in saving prf
- Item Masterlist format: ID | ITEM NAME | DESCRIPTION to ITEM NAME | DESCRIPTION | SUPPLIER

* Changed available to item unit in item field list.

* Added inputmask for contact number and added telephone field in customers table.
Added validation format for email and contact number.

* Adjested the constraint from 20 to 50

* Added a checkbox where user can check if client does not have a contact number or telephone number only.

* Minor codes changes

+++ app/Controllers/Settings/GeneralInfo.php
@@ -162,9 +162,11 @@ class GeneralInfo extends BaseController
                     return $data;
                 }
 
-                // Remove the previous file
-                $filepath = $this->fullFilePathLogo() . $curFilename;
-                $this->removeFile($filepath);
+                if (! empty($curFilename)) {
+                    // Remove the previous file
+                    $filepath = $this->fullFilePathLogo() . $curFilename;
+                    $this->removeFile($filepath);
+                }
                 
                 $data['files'] = $file;
                 return $data;
@@ -192,10 +194,14 @@ class GeneralInfo extends BaseController
                 if ($q = $this->request->getVar('q')) {
                     $files              = [];
                     $filename           = $this->getGeneralInfo($q);
-                    $nfilename          = strpos($filename, '/') ? explode('/', $filename) : $filename;
-                    $nfilename          = is_array($nfilename) ? array_pop($nfilename) : $nfilename;
-                    $downloadUrl        = base_url($this->initialFilePathLogo . $nfilename);
-                    $files              = $this->getFiles($q, [$nfilename], $this->fullFilePathLogo(), $downloadUrl);
+
+                    if (! empty($filename)) {
+                        $nfilename          = strpos($filename, '/') ? explode('/', $filename) : $filename;
+                        $nfilename          = is_array($nfilename) ? array_pop($nfilename) : $nfilename;
+                        $downloadUrl        = base_url($this->initialFilePathLogo . $nfilename);
+                        $files              = $this->getFiles($q, [$nfilename], $this->fullFilePathLogo(), $downloadUrl);
+                    }
+
                     $data['files']      = $files;
                 } else {
                     $data['data']       = $this->_model->fetchAll();
